Add files via upload
Some new versions of firmware for the original Nano have broken from the 101 data point standard.  This version allows the user to change the number of data points, similar to the V2+ software.  This setting is stored into the defaults file.  It is up to the user to know how many data points their firmware supports.     In the past, this was calculated by sending the frequency command.  This software no longer uses that command to help improve the performance.